update(domains): Update domainsets for loop
using letsencrypt.get('domainsets', {}).items() will prevent the state from failing if no domainsets are defined. 

This is particularly useful for environments where letsencrypt can be included in the top file for a group of servers where only a few of them have domainsets defined. The servers that don't have domainsets defined will not fail to run scheduled highstates.